+++ azext_iot/central/providers/job_provider.py
@@ -5,13 +5,12 @@
 # --------------------------------------------------------------------------------------------
 
 
-from typing import List, Union
+from typing import List
 from knack.log import get_logger
 from azure.cli.core.azclierror import AzureResponseError, ClientRequestError, ResourceNotFoundError
 from azext_iot.constants import CENTRAL_ENDPOINT
 from azext_iot.central import services as central_services
-from azext_iot.central.models.preview import JobPreview
-from azext_iot.central.models.v1_1_preview import JobV1_1_preview
+from azext_iot.central.models.ga_2022_07_31 import JobGa
 
 logger = get_logger(__name__)
 
@@ -38,7 +37,7 @@ class CentralJobProvider:
 
     def list_jobs(
         self, central_dns_suffix=CENTRAL_ENDPOINT
-    ) -> List[Union[JobPreview, JobV1_1_preview]]:
+    ) -> List[JobGa]:
         jobs = central_services.job.list_jobs(
             cmd=self._cmd,
             app_id=self._app_id,
@@ -56,7 +55,7 @@ class CentralJobProvider:
         self,
         job_id,
         central_dns_suffix=CENTRAL_ENDPOINT,
-    ) -> Union[JobPreview, JobV1_1_preview]:
+    ) -> JobGa:
         # get or add to cache
         job = self._jobs.get(job_id)
         if not job:
@@ -79,7 +78,7 @@ class CentralJobProvider:
         self,
         job_id,
         central_dns_suffix=CENTRAL_ENDPOINT,
-    ) -> Union[JobPreview, JobV1_1_preview]:
+    ) -> JobGa:
         # get or add to cache
         job = central_services.job.stop_job(
             cmd=self._cmd,
@@ -98,7 +97,7 @@ class CentralJobProvider:
         self,
         job_id,
         central_dns_suffix=CENTRAL_ENDPOINT,
-    ) -> Union[JobPreview, JobV1_1_preview]:
+    ) -> JobGa:
         # get or add to cache
         job = central_services.job.resume_job(
             cmd=self._cmd,
@@ -119,7 +118,7 @@ class CentralJobProvider:
         job_id,
         rerun_id,
         central_dns_suffix=CENTRAL_ENDPOINT,
-    ) -> Union[JobPreview, JobV1_1_preview]:
+    ) -> JobGa:
         # get or add to cache
         job = central_services.job.rerun_job(
             cmd=self._cmd,